@@ -1,13 +1,12 @@
 #ifndef MAYBSFML_CONFIG_H
 #define MAYBSFML_CONFIG_H
 
-
-struct Config {
-    const unsigned int width        = 1260;
-    const unsigned int height       = 770;
-    const unsigned int fps_cap      = 120;
-    const bool         isVSyncOn    = true;
+struct Config
+{
+    const unsigned int width = 1260;
+    const unsigned int height = 770;
+    const unsigned int fps_cap = 120;
+    const bool isVSyncOn = true;
 };
 
-
-#endif //MAYBSFML_CONFIG_H
+#endif  // MAYBSFML_CONFIG_H
