@@ -1,45 +1,62 @@
 #include "Entity.hpp"
 
 Entity::Entity(sf::Vector2i size, sf::Vector2f position, sf::Vector2i offset)
-        : m_size(size), m_position(position), m_velocity({0.f, 0.f}),
-          m_bounds(m_position, size), grounded(false),
-          Renderable(offset, size), Serializable() {
-    m_sprite.setTextureRect(m_indicator);
-    m_sprite.setPosition(position);
+    : size_(size)
+    , position_(position)
+    , velocity_({0.f, 0.f})
+    , bounds_(position_, size)
+    , grounded(false)
+    , Renderable(offset, size)
+    , Serializable()
+{
+    sprite_.setTextureRect(indicator_);
+    sprite_.setPosition(position);
 }
 
-void Entity::resolveCollision(CollisionType collType) {
-    if (collType == CollisionType::Vertical) {
-        m_velocity.x = 0;
-        grounded = false;
-    } else if (collType == CollisionType::Horizontal || collType == CollisionType::HorizontalGround) {
-        m_velocity.y = 0;
-    } else if (collType == CollisionType::Both || collType == CollisionType::BothGround) {
-        m_velocity = {0, 0};
+void Entity::resolveCollision(CollisionType collType)
+{
+    if (collType == CollisionType::Vertical)
+    {
+        velocity_.x = 0;
         grounded = false;
     }
-    if (collType == CollisionType::HorizontalGround || collType == CollisionType::BothGround) {
+    else if (collType == CollisionType::Horizontal
+             || collType == CollisionType::HorizontalGround)
+    {
+        velocity_.y = 0;
+    }
+    else if (collType == CollisionType::Both || collType == CollisionType::BothGround)
+    {
+        velocity_ = {0, 0};
+        grounded = false;
+    }
+    if (collType == CollisionType::HorizontalGround || collType == CollisionType::BothGround)
+    {
         grounded = true;
     }
 }
 
-const sf::Vector2f& Entity::velocity() const {
-    return m_velocity;
+const sf::Vector2f& Entity::velocity() const
+{
+    return velocity_;
 }
 
-Collider& Entity::AABBox() {
-    return m_bounds;
+Collider& Entity::AABBox()
+{
+    return bounds_;
 }
 
-void Entity::writeObject() {
+void Entity::writeObject()
+{
     std::ofstream f(filepath + "Entity.ser", f.binary);
-    write(f, m_position);
-    write(f, m_velocity);
-    write(f, m_size);
+    write(f, position_);
+    write(f, velocity_);
+    write(f, size_);
     f.close();
 }
 
-void Entity::readObject() {
+void Entity::readObject()
+{
     sf::Vector2f pos;
     sf::Vector2f vel;
     sf::Vector2i siz;
