@@ -1,21 +1,26 @@
 #ifndef MAYBSFML_ENTITY_HPP
 #define MAYBSFML_ENTITY_HPP
 
-
+#include "Maths/Collider.hpp"
 #include "Rendering/Renderable.hpp"
 #include "States/Serializable.hpp"
-#include "Maths/Collider.hpp"
 
+class Entity
+    : public Renderable
+    , public Serializable
+{
+    sf::Vector2i size_;
+    Collider bounds_;
 
-class Entity : public Renderable, public Serializable {
-    sf::Vector2i    m_size;
-    Collider        m_bounds;
 protected:
-    sf::Vector2f    m_velocity;
-    sf::Vector2f    m_position;
-    bool            grounded;
+    sf::Vector2f velocity_;
+    sf::Vector2f position_;
+    bool grounded;
+
 public:
-    explicit Entity(sf::Vector2i size, sf::Vector2f position={0, 0}, sf::Vector2i offset={0, 0});
+    explicit Entity(sf::Vector2i size,
+                    sf::Vector2f position = {0, 0},
+                    sf::Vector2i offset = {0, 0});
 
     void resolveCollision(CollisionType);
 
@@ -27,11 +32,11 @@ public:
 
     void readObject() override;
 
-    ~Entity() override {
-//        writeObject();
-//        readObject();
+    ~Entity() override
+    {
+        //        writeObject();
+        //        readObject();
     }
 };
 
-
-#endif //MAYBSFML_ENTITY_HPP
+#endif  // MAYBSFML_ENTITY_HPP
