@@ -1,70 +1,87 @@
-#include "Player.hpp"
-#include "../TextureLoader.hpp"
+#include "Player/Player.hpp"
+
+#include "TextureCache.hpp"
 
 Player::Player(sf::Vector2f position)
-        : Entity{{80, 110}, position}, m_animCounter(0) {
-    AABBox().setMargin(-5, 0, 0 ,0);
-    auto& tex = TextureLoader::instance();
-    m_texture = tex.add("Player", "../Resources/World/player.png");
-    m_texture_mirror = tex.add("PlayerMirror", "../Resources/World/player_mirror.png");
-    m_sprite.setTexture(*m_texture);
-    m_sprite.setTextureRect(m_indicator);
+    : Entity{{80, 110}, position}
+    , animCounter_(0)
+{
+    auto&& cache = TextureCache::getInstance();
+    texture_ = &cache.fetch("Player", std_fs::path("World") / "player.png");
+    texture_mirror_ = &cache.fetch("PlayerMirror",
+                                   std_fs::path("World") / "player_mirror.png");
+    AABBox().setMargin(-5, 0, 0, 0);
+    sprite_.setTexture(*texture_);
+    sprite_.setTextureRect(indicator_);
 }
 
-void Player::handleKeyboardInput(float dt) {
-    m_velocity.x = 0;
-    m_indicator.left = 0;
-    m_indicator.top = 0;
-    if (sf::Keyboard::isKeyPressed(sf::Keyboard::A)) {
-        m_velocity.x -= speed * dt;
-        m_sprite.setTexture(*m_texture_mirror);
-        m_indicator.top = 110;
-        m_indicator.left = 80;
-        if (m_animCounter > 7) {
-            m_indicator.left = (m_indicator.left + 80) % 160;
+void Player::handleKeyboardInput(float dt)
+{
+    velocity_.x = 0;
+    indicator_.left = 0;
+    indicator_.top = 0;
+    if (sf::Keyboard::isKeyPressed(sf::Keyboard::A))
+    {
+        velocity_.x -= speed * dt;
+        sprite_.setTexture(*texture_mirror_);
+        indicator_.top = 110;
+        indicator_.left = 80;
+        if (animCounter_ > 7)
+        {
+            indicator_.left = (indicator_.left + 80) % 160;
         }
     }
-    if (sf::Keyboard::isKeyPressed(sf::Keyboard::D)) {
-        m_velocity.x += speed * dt;
-        m_sprite.setTexture(*m_texture);
-        m_indicator.top = 110;
-        if (m_animCounter > animDelay / 2) {
-            m_indicator.left = (m_indicator.left + 80) % 160;
+    if (sf::Keyboard::isKeyPressed(sf::Keyboard::D))
+    {
+        velocity_.x += speed * dt;
+        sprite_.setTexture(*texture_);
+        indicator_.top = 110;
+        if (animCounter_ > animDelay / 2)
+        {
+            indicator_.left = (indicator_.left + 80) % 160;
         }
     }
-    m_animCounter = (m_animCounter + 1) % animDelay;
-    if (sf::Keyboard::isKeyPressed(sf::Keyboard::W) and grounded and m_velocity.y == 0) {
-        m_velocity.y = -20.f;
+    animCounter_ = (animCounter_ + 1) % animDelay;
+    if (sf::Keyboard::isKeyPressed(sf::Keyboard::W) and grounded and velocity_.y == 0)
+    {
+        velocity_.y = -20.f;
         grounded = false;
     }
 }
 
-void Player::update(float dt) {
+void Player::update(float dt)
+{
     handleKeyboardInput(dt);
     // Apply gravity
-    m_velocity.y += 1.f;
+    velocity_.y += 1.f;
 }
 
-void Player::move() {
-    if (not grounded and m_velocity.y != 0) {
-        m_indicator.top = 0;
-        if (m_velocity.y < 0.f) m_indicator.left = 80;
-        else m_indicator.left = 160;
+void Player::move()
+{
+    if (not grounded and velocity_.y != 0)
+    {
+        indicator_.top = 0;
+        if (velocity_.y < 0.f)
+            indicator_.left = 80;
+        else
+            indicator_.left = 160;
     }
-    m_position.x += m_velocity.x;
-    m_position.y += m_velocity.y;
-    m_sprite.setPosition(m_position);
-    m_sprite.setTextureRect(m_indicator);
+    position_.x += velocity_.x;
+    position_.y += velocity_.y;
+    sprite_.setPosition(position_);
+    sprite_.setTextureRect(indicator_);
 }
 
-void Player::writeObject() {
+void Player::writeObject()
+{
     std::ofstream f(filepath + "Player.ser", f.binary);
-    write(f, m_position);
-    write(f, m_velocity);
+    write(f, position_);
+    write(f, velocity_);
     f.close();
 }
 
-void Player::readObject() {
+void Player::readObject()
+{
     sf::Vector2f pos;
     sf::Vector2f vel;
     std::ifstream f(filepath + "Player.ser", f.binary);
