@@ -1,27 +1,38 @@
 #include "RenderManager.hpp"
 
-RenderManager::RenderManager(sf::RenderWindow& window) : m_window(window) {}
-
-void RenderManager::drawSFML(const sf::Drawable& drawable) const {
-    m_window.draw(drawable);
+RenderManager::RenderManager(sf::RenderWindow& window)
+    : window_(window)
+{
 }
 
-void RenderManager::drawSFML(const sf::Vertex* vertices, size_t vertexCount, sf::PrimitiveType p) const {
-    m_window.draw(vertices, vertexCount, p);
+void RenderManager::drawSFML(const sf::Drawable& drawable) const
+{
+    window_.draw(drawable);
 }
 
-void RenderManager::drawBlock(const Block& block) const {
-    m_window.draw(block.sprite());
+void RenderManager::drawSFML(const sf::Vertex* vertices,
+                             size_t vertexCount,
+                             sf::PrimitiveType p) const
+{
+    window_.draw(vertices, vertexCount, p);
 }
 
-void RenderManager::drawEntity(Entity& entity) const {
-    m_window.draw(entity.sprite());
+void RenderManager::drawBlock(const Block& block) const
+{
+    window_.draw(block.sprite());
 }
 
-void RenderManager::loadSprite(const sf::Sprite& s) const {
-    m_window.draw(s);
+void RenderManager::drawEntity(Entity& entity) const
+{
+    window_.draw(entity.sprite());
 }
 
-void RenderManager::refresh(sf::Color color) const {
-    m_window.clear(color);
+void RenderManager::loadSprite(const sf::Sprite& s) const
+{
+    window_.draw(s);
+}
+
+void RenderManager::refresh(sf::Color color) const
+{
+    window_.clear(color);
 }
