@@ -1,14 +1,14 @@
 #ifndef MAYBSFML_RENDERMANAGER_HPP
 #define MAYBSFML_RENDERMANAGER_HPP
 
-
-#include <SFML/Graphics.hpp>
-#include "../World/Objects/Block.hpp"
 #include "../Entity.hpp"
+#include "../World/Objects/Block.hpp"
+#include <SFML/Graphics.hpp>
 
+class RenderManager
+{
+    sf::RenderWindow& window_;
 
-class RenderManager {
-    sf::RenderWindow& m_window;
 public:
     explicit RenderManager(sf::RenderWindow& window);
 
@@ -18,12 +18,11 @@ public:
 
     void drawEntity(Entity& entity) const;
 
-    void loadSprite(const sf::Sprite &) const;
+    void loadSprite(const sf::Sprite&) const;
 
     void refresh(sf::Color color) const;
 
     void drawSFML(const sf::Vertex* vertices, size_t vertexCount, sf::PrimitiveType p) const;
 };
 
-
-#endif //MAYBSFML_RENDERMANAGER_HPP
+#endif  // MAYBSFML_RENDERMANAGER_HPP
