@@ -1,42 +1,57 @@
 #include "GameScene.hpp"
 
 GameScene::GameScene(const Config& config, const FPSCounter& fpsCounter)
-        : m_world(config), m_counter(fpsCounter), popUpMenu() {
+    : world_(config)
+    , counter_(fpsCounter)
+    , popUpMenu()
+{
 }
 
-void GameScene::update() {
-    if (not popUpMenu.isOpen()) {
-        m_world.update(m_counter.frametime());
+void GameScene::update()
+{
+    if (not popUpMenu.isOpen())
+    {
+        world_.update(counter_.frametime());
     }
-    if (m_world.isGameOver()) {
-        hasFinished = true;
+    if (world_.isGameOver())
+    {
+        endScene();
     }
 }
 
-void GameScene::render(const RenderManager& renderer) {
-    m_world.render(renderer);
-    if (popUpMenu.isOpen()) {
+void GameScene::render(const RenderManager& renderer)
+{
+    world_.render(renderer);
+    if (popUpMenu.isOpen())
+    {
         popUpMenu.render(renderer);
     }
     // TODO Possible HUD rendering here...
 }
 
-const std::string GameScene::nextScene() const {
+std::string_view GameScene::nextScene() const
+{
     return "MenuScene";
 }
 
-void GameScene::handleEvents(sf::RenderWindow& window) {
-    if (popUpMenu.isOpen()) {
-        popUpMenu.update(window, hasFinished);
+void GameScene::handleEvents(sf::RenderWindow& window)
+{
+    if (popUpMenu.isOpen())
+    {
+        popUpMenu.update(window, *this);
         return;
     }
     sf::Event event{};
-    while (window.pollEvent(event)) {
-        if (event.type == sf::Event::Closed) {
+    while (window.pollEvent(event))
+    {
+        if (event.type == sf::Event::Closed)
+        {
             window.close();
         }
-        if (event.type == sf::Event::KeyPressed) {
-            if (event.key.code == sf::Keyboard::Escape) {
+        if (event.type == sf::Event::KeyPressed)
+        {
+            if (event.key.code == sf::Keyboard::Escape)
+            {
                 popUpMenu.toggle();
             }
         }
