@@ -1,25 +1,28 @@
 #ifndef MAYBSFML_SERIALIZABLE_HPP
 #define MAYBSFML_SERIALIZABLE_HPP
 
-
 #include <SFML/System/Vector2.hpp>
 #include <cstdio>
-#include <string>
 #include <fstream>
+#include <string>
 
-class Serializable {
+class Serializable
+{
 protected:
     const std::string filepath = "../Saves/";
 
-    template<typename T>
-    void read(std::ifstream& file, T& var) {
+    template <typename T>
+    void read(std::ifstream& file, T& var)
+    {
         file.read(reinterpret_cast<char*>(&var), sizeof var);
     }
 
-    template<typename T>
-    void write(std::ofstream& file, T& var) {
+    template <typename T>
+    void write(std::ofstream& file, T& var)
+    {
         file.write(reinterpret_cast<char*>(&var), sizeof var);
     }
+
 public:
     explicit Serializable() = default;
 
@@ -29,5 +32,4 @@ public:
     virtual ~Serializable() = default;
 };
 
-
-#endif //MAYBSFML_SERIALIZABLE_HPP
+#endif  // MAYBSFML_SERIALIZABLE_HPP
