@@ -1,35 +1,45 @@
-#include "FPSCounter.hpp"
+#include "Utils/FPSCounter.hpp"
+
+#include "Utils/ResourceManager.hpp"
+
+#include <iostream>
 
 FPSCounter::FPSCounter()
-        : m_fpsTimer(), m_currentFrames(0), m_fps(0), m_tickTimer(), m_lastTick(0) {
-    m_text.move(10, 10);
-    m_text.setOutlineColor(sf::Color::Blue);
-    m_text.setOutlineThickness(2);
-    m_font.loadFromFile("../Resources/Fonts/arial_bold.ttf");
-    m_text.setFont(m_font);
-    m_text.setCharacterSize(25);
+    : currentFrames_(0)
+    , fps_(0)
+    , lastTick_(0)
+    , font_(ResourceManager::load<sf::Font>(std_fs::path("Fonts") / "arial_bold.ttf"))
+{
+    text_.move(sf::Vector2f(10, 10));
+    text_.setOutlineColor(sf::Color::Blue);
+    text_.setOutlineThickness(2);
+    text_.setFont(font_);
+    text_.setCharacterSize(25);
 }
 
-float slotTime = .5f;
+void FPSCounter::update()
+{
+    constexpr float slotTime = .5f;
+    ++currentFrames_;
 
-void FPSCounter::update() {
-    m_currentFrames++;
-
-    if (m_fpsTimer.getElapsedTime().asSeconds() > slotTime) {
-        m_fps = m_currentFrames / m_fpsTimer.getElapsedTime().asSeconds();
-        m_fpsTimer.restart();
-        m_currentFrames = 0;
+    if (fpsTimer_.getElapsedTime().asSeconds() > slotTime)
+    {
+        fps_ = currentFrames_ / fpsTimer_.getElapsedTime().asSeconds();
+        fpsTimer_.restart();
+        currentFrames_ = 0;
     }
-    m_lastTick = m_tickTimer.restart().asSeconds();
+    lastTick_ = tickTimer_.restart().asSeconds();
 }
 
-void FPSCounter::draw(const RenderManager& renderer) {
+void FPSCounter::draw(const RenderManager& renderer)
+{
     std::stringstream ss;
-    ss << std::fixed << std::setprecision(2) << m_fps;
-    m_text.setString("FPS: " + ss.str());
-    renderer.drawSFML(m_text);
+    ss << std::fixed << std::setprecision(2) << fps_;
+    text_.setString("FPS: " + ss.str());
+    renderer.drawSFML(text_);
 }
 
-float FPSCounter::frametime() const {
-    return m_lastTick;
+float FPSCounter::frametime() const
+{
+    return lastTick_;
 }
