@@ -1,10 +1,8 @@
 #ifndef MAYBSFML_GAMESAVE_HPP
 #define MAYBSFML_GAMESAVE_HPP
 
-
-class GameSave {
-
+class GameSave
+{
 };
 
-
-#endif //MAYBSFML_GAMESAVE_HPP
+#endif  // MAYBSFML_GAMESAVE_HPP
