@@ -1,20 +1,22 @@
 #ifndef MAYBSFML_BUTTONS_HPP
 #define MAYBSFML_BUTTONS_HPP
 
+#include "Utils/ResourceManager.hpp"
 
-#include <SFML/Graphics/Texture.hpp>
 #include <SFML/Graphics/Sprite.hpp>
+#include <SFML/Graphics/Texture.hpp>
 
-class SpriteBuilder {
-    std::vector<sf::Sprite>     m_sprites;
-    std::vector<sf::Vector2f>   m_positions;
+class SpriteBuilder
+{
+    std::vector<sf::Sprite> sprites_;
+    std::vector<sf::Vector2f> positions_;
+
 public:
-    SpriteBuilder();
+    SpriteBuilder() = default;
 
-    void build(std::string imagepath, const sf::Vector2f& position);
+    void build(const std_fs::path& imagepath, const sf::Vector2f& position);
 
     const std::vector<sf::Sprite>& sprites() const;
 };
 
-
-#endif //MAYBSFML_BUTTONS_HPP
+#endif  // MAYBSFML_BUTTONS_HPP
