@@ -1,32 +1,32 @@
 #ifndef MAYBSFML_TEXTURESTASH_HPP
 #define MAYBSFML_TEXTURESTASH_HPP
 
+#include "TextureCache.hpp"
 
-#include <SFML/Graphics/Texture.hpp>
 #include <SFML/Graphics/Sprite.hpp>
+#include <SFML/Graphics/Texture.hpp>
+
 #include <map>
-#include "../TextureLoader.hpp"
 
-class Stash {
-    sf::Texture*                        m_tileset;
-    std::map<std::string, sf::IntRect>  m_tileTypeMap;
+class Stash
+{
+    const sf::Texture& tileset_ =
+        TextureCache::getInstance().fetch("Stash.tileset",
+                                          std_fs::path("World") / "tileset.png");
+    std::map<std::string, sf::IntRect> tileTypeMap_;
+
 public:
-    Stash() {
-        m_tileset = TextureLoader::instance().add("", "../Resources/World/tileset.png");
+    void add(const std::string& tileTypeName, int x, int y)
+    {
+        tileTypeMap_.insert({tileTypeName, sf::IntRect(x, y, 70, 70)});
     }
 
-    void add(const std::string& tileTypeName, int x, int y) {
-        m_tileTypeMap.insert({tileTypeName, sf::IntRect(x, y, 70, 70)});
+    sf::Sprite getSprite(const std::string& tileTypeName) const
+    {
+        return {tileset_, tileTypeMap_.find(tileTypeName)->second};
     }
 
-    sf::Sprite getSprite(const std::string& tileTypeName) const {
-        return sf::Sprite(*m_tileset, m_tileTypeMap.find(tileTypeName)->second);
-    }
-
-    const auto& data() const {
-        return m_tileTypeMap;
-    }
+    const auto& data() const { return tileTypeMap_; }
 };
 
-
-#endif //MAYBSFML_TEXTURESTASH_HPP
+#endif  // MAYBSFML_TEXTURESTASH_HPP
