@@ -1,17 +1,24 @@
 #include "Block.hpp"
 
-Block::Block(const sf::Texture& tiles, sf::Vector2f position, sf::Vector2i size, sf::Vector2i offset)
-        : GameObject(position, size, offset), m_bounds(m_position, size),  m_sprite() {
+Block::Block(const sf::Texture& tiles,
+             sf::Vector2f position,
+             sf::Vector2i size,
+             sf::Vector2i offset)
+    : GameObject(position, size, offset)
+    , bounds_(position_, size)
+{
 
-    m_sprite.setTexture(tiles);
-    m_sprite.setTextureRect(m_indicator);
-    m_sprite.setPosition(position);
+    sprite_.setTexture(tiles);
+    sprite_.setTextureRect(indicator_);
+    sprite_.setPosition(position);
 }
 
-Collider Block::AABBox() const {
-    return m_bounds;
+Collider Block::AABBox() const
+{
+    return bounds_;
 }
 
-const sf::Sprite& Block::sprite() const {
-    return m_sprite;
+const sf::Sprite& Block::sprite() const
+{
+    return sprite_;
 }
