@@ -1,14 +1,16 @@
 #ifndef MAYBSFML_GAMEOBJECT_HPP
 #define MAYBSFML_GAMEOBJECT_HPP
 
-
 #include <SFML/Graphics.hpp>
 
-class GameObject {
-    sf::Vector2i        m_size;
+class GameObject
+{
+    sf::Vector2i size_;
+
 protected:
-    sf::Vector2f        m_position;
-    sf::IntRect         m_indicator;
+    sf::Vector2f position_;
+    sf::IntRect indicator_;
+
 public:
     explicit GameObject(sf::Vector2f position, sf::Vector2i size, sf::Vector2i offset);
 
@@ -17,5 +19,4 @@ public:
     virtual ~GameObject() = default;
 };
 
-
-#endif //MAYBSFML_GAMEOBJECT_HPP
+#endif  // MAYBSFML_GAMEOBJECT_HPP
