@@ -1,25 +1,27 @@
 #ifndef MAYBSFML_WORLDLOADER_HPP
 #define MAYBSFML_WORLDLOADER_HPP
 
+#include "Config.hpp"
+#include "Entity.hpp"
+#include "Objects/Block.hpp"
+#include "Player/Player.hpp"
+#include "Utils/ResourceManager.hpp"
 
 #include <vector>
-#include "Objects/Block.hpp"
-#include "../Player/Player.hpp"
-#include "../Config.hpp"
-#include "../Entity.hpp"
 
-class WorldLoader {
+class WorldLoader
+{
     const Config& config;
+
 protected:
-    std::vector<std::shared_ptr<Block>>     m_blocks;
-    std::vector<std::shared_ptr<Entity>>    m_entities;
+    std::vector<std::shared_ptr<Block>> blocks_;
+    std::vector<std::shared_ptr<Entity>> entities_;
+
 public:
     explicit WorldLoader(const Config&);
 
-    void load(std::string filepath, std::shared_ptr<Player>&);
-
-    void reload(std::string filepath);
+    void load(const std_fs::path& filepath, std::shared_ptr<Player>&);
+    void reload(const std_fs::path& filepath);
 };
 
-
-#endif //MAYBSFML_WORLDLOADER_HPP
+#endif  // MAYBSFML_WORLDLOADER_HPP
