@@ -1,13 +1,13 @@
 #include "NoFunction.hpp"
 
-void NoFunction::render(const RenderManager& renderer) {
-
+void NoFunction::render(const RenderManager& renderer)
+{
 }
 
-void NoFunction::mousePressed(int posx, int posy) {
-
+void NoFunction::mousePressed(int posx, int posy)
+{
 }
 
-void NoFunction::mouseMoved(int posx, int posy) {
-
+void NoFunction::mouseMoved(int posx, int posy)
+{
 }
