@@ -1,18 +1,16 @@
 #ifndef MAYBSFML_NOFUNCTION_HPP
 #define MAYBSFML_NOFUNCTION_HPP
 
-
-#include <SFML/Graphics.hpp>
-#include "IFunctionState.hpp"
 #include "../../Rendering/RenderManager.hpp"
+#include "IFunctionState.hpp"
+#include <SFML/Graphics.hpp>
 
 class FunctionManager;
 
-class NoFunction : public IFunctionState {
+class NoFunction : public IFunctionState
+{
 public:
-    explicit NoFunction(FunctionManager& fmref) {
-        printf("NoFunction\n");
-    };
+    explicit NoFunction(FunctionManager& fmref) { printf("NoFunction\n"); };
 
     void render(const RenderManager& renderer) override;
 
@@ -21,5 +19,4 @@ public:
     void mouseMoved(int posx, int posy) override;
 };
 
-
-#endif //MAYBSFML_NOFUNCTION_HPP
+#endif  // MAYBSFML_NOFUNCTION_HPP
