@@ -1,15 +1,16 @@
 #ifndef MAYBSFML_REMOVALFUNCTION_HPP
 #define MAYBSFML_REMOVALFUNCTION_HPP
 
-
 #include "IFunctionState.hpp"
 
 class FunctionManager;
 
-class RemovalFunction : public IFunctionState {
-    FunctionManager&    m_manager;
-    sf::RectangleShape  m_indicator;
-    sf::Vector2i        m_lastMousePos;
+class RemovalFunction : public IFunctionState
+{
+    FunctionManager& manager_;
+    sf::RectangleShape indicator_;
+    sf::Vector2i lastMousePos_;
+
 public:
     explicit RemovalFunction(FunctionManager& fmref);
 
@@ -20,5 +21,4 @@ public:
     void mouseMoved(int posx, int posy) override;
 };
 
-
-#endif //MAYBSFML_REMOVALFUNCTION_HPP
+#endif  // MAYBSFML_REMOVALFUNCTION_HPP
